Binding user to cart when QRCode read
